feat: Setup do flyway para migrations. Ajustes do código para suportar as migrations.

+++ src/main/java/com/arthwh/registroReceitas/dto/ReceitaRegisterDTO.java
@@ -2,10 +2,12 @@ package com.arthwh.registroReceitas.dto;
 
 import com.arthwh.registroReceitas.model.TipoReceitaEnum;
 
+import java.math.BigDecimal;
+
 public record ReceitaRegisterDTO(
         String nome,
         String descricao,
-        Double custo,
+        BigDecimal custo,
         TipoReceitaEnum tipoReceita
 ) {
 }
