+++ src/main/java/com/arthwh/registroReceitas/dto/ReceitaUpdateDTO.java
@@ -1,8 +1,10 @@
 package com.arthwh.registroReceitas.dto;
 
+import java.math.BigDecimal;
+
 public record ReceitaUpdateDTO(
         Integer id,
         String descricao,
-        Double custo
+        BigDecimal custo
 ) {
 }
